+++ src/components/MyKiva/LendingStats.vue
@@ -1,11 +1,8 @@
 <template>
 	<div class="tw-mb-2">
 		<h3 class="tw-text-primary tw-mb-1">
-			{{ title }}
+			Next steps recommended for you
 		</h3>
-		<p class="tw-text-base">
-			{{ description }}
-		</p>
 	</div>
 	<div
 		ref="loanRegionsElement"
@@ -32,6 +29,7 @@
 					:categories-loan-count="categoriesLoanCount"
 					:hide-goal-card="hideCompletedGoalCard"
 					:post-lending-next-steps-enable="postLendingNextStepsEnable"
+					:user-info="userInfo"
 					@open-goal-modal="showGoalModal = true"
 				/>
 			</div>
@@ -162,6 +160,7 @@
 			:hide-goal-card="hideCompletedGoalCard"
 			:post-lending-next-steps-enable="postLendingNextStepsEnable"
 			:latest-loan="latestLoan"
+			:user-info="userInfo"
 			@open-goal-modal="showGoalModal = true"
 		/>
 		<GoalSettingModal
@@ -267,7 +266,11 @@ export default {
 		goalRefreshKey: {
 			type: Number,
 			default: 0
-		}
+		},
+		userInfo: {
+			type: Object,
+			default: () => ({}),
+		},
 	},
 	data() {
 		return {
@@ -293,9 +296,6 @@ export default {
 		loanRegions() {
 			return this.regionsData.filter(region => region.hasLoans).length;
 		},
-		hasLoans() {
-			return this.loans.length > 0;
-		},
 		pillHeader() {
 			if (this.totalRegions === 0) return '';
 			if (this.loanRegions === 0) return 'Make a global impact';
@@ -319,16 +319,6 @@ export default {
 		categoriesLoanCount() {
 			const { getAllCategoryLoanCounts } = useBadgeData();
 			return getAllCategoryLoanCounts(this.heroTieredAchievements);
-		},
-		title() {
-			if (!this.hasLoans) return 'Your impact starts here';
-			if (this.isNextStepsExpEnabled) return 'Make a difference today';
-			return 'Ready to grow your impact?';
-		},
-		description() {
-			if (!this.hasLoans) return 'Recommended for you';
-			if (this.isNextStepsExpEnabled) return 'How many more people will you help this year?';
-			return 'Next steps for you based on your lending history';
 		},
 	},
 	setup() {